+++ CHANGES.rst
@@ -1,3 +1,6 @@
+0.3.2
+-----
+
 0.3.1
 -----
 
@@ -24,6 +27,8 @@
 * Optional support for Flask-DebugToolbar and LineProfiler extension
 * SelectField now supports optgroup grouping
 * Comparison validators to complement ``EqualTo`` with ``>``, ``>=``, ``<``, ``<=``, ``!=``
+* DateTimeField's timezone parameter now defaults to user timezone, not UTC
+* New form filters to replace the (overloaded) use of validators to filter
 
 0.3.0
 -----